Crate qutex

A collection of locking data structures, both thread-safe and single-thread-optimized, which use Rust futures instead of thread-blocking.

Structs

FutureGuard
A future which resolves to a Guard.

FutureReadGuard
A future which resolves to a ReadGuard.

FutureWriteGuard
A future which resolves to a WriteGuard.

Guard
Allows access to the data contained within a lock just like a mutex guard.

QrwLock
A queue-backed read/write data lock.

QrwRequest
A request to lock the lock for either read or write access.

Qutex
A lock-free-queue-backed exclusive data lock.

ReadGuard
Allows read-only access to the data contained within a lock.

Request
A request to lock the qutex for exclusive access.

WriteGuard
Allows read or write access to the data contained within a lock.

Enums

RequestKind
Specifies whether a QrwRequest is a read or write request.