Struct qutex::Guard

pub struct Guard<T> { /* fields omitted */ }

Allows access to the data contained within a lock just like a mutex guard.

Methods

impl<T> Guard<T>

pub fn unlock(guard: Guard<T>) -> Qutex<T>

Releases the lock held by a Guard and returns the original Qutex.