Struct qutex::QrwLock

pub struct QrwLock<T> { /* fields omitted */ }

A queue-backed read/write data lock.

As with any queue-backed system, deadlocks must be carefully avoided when interoperating with other queues.

Methods

impl<T> QrwLock<T>

pub fn new(val: T) -> QrwLock<T>

Creates and returns a new QrwLock.

pub fn read(self) -> FutureReadGuard<T>

Returns a new FutureReadGuard which can be used as a future and will resolve into a ReadGuard.

pub fn write(self) -> FutureWriteGuard<T>

Returns a new FutureWriteGuard which can be used as a future and will resolve into a WriteGuard.

pub unsafe fn enqueue_lock_request(&self, req: QrwRequest)

Pushes a lock request onto the queue.

pub fn get_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the inner Vec if there are currently no other copies of this QrwLock.

Since this call borrows the inner lock mutably, no actual locking needs to take place---the mutable borrow statically guarantees no locks exist.

pub fn as_ptr(&self) -> *const T

Returns a reference to the inner value.

pub fn as_mut_ptr(&self) -> *mut T

Returns a mutable reference to the inner value.

pub unsafe fn process_queues(&self)

Pops the next lock request in the queue if possible.

pub unsafe fn upgrade_read_lock(&self) -> Result<(), Receiver<()>>

Converts a single read lock (read count of '1') into a write lock.

Returns an error containing a oneshot receiver if there is currently more than one read lock. When the read count reaches one, the receiver channel will be completed (i.e. poll it).

Panics if there are no read locks.

Do not call this method directly unless you are using a custom guard or are otherwise managing the lock state manually. Use ReadGuard::upgrade instead.

pub unsafe fn downgrade_write_lock(&self)

Converts a write lock into a read lock then processes the queue, allowing additional read requests to acquire locks.

Use WriteGuard::downgrade rather than calling this directly.

pub unsafe fn release_read_lock(&self)

Decreases the reader count by one and unparks the next requester task in the queue if possible.

If a reader is waiting to be upgraded and the read lock count reaches 1, the upgrade sender will be completed.

pub unsafe fn release_write_lock(&self)

Unlocks this (the caller's) lock and unparks the next requester task in the queue if possible.